Move item views to own file
