+++ app/stac_api/urls.py
@@ -6,6 +6,8 @@ from rest_framework.authtoken.views import obtain_auth_token
 
 from stac_api.views.collection import CollectionDetail
 from stac_api.views.collection import CollectionList
+from stac_api.views.item import ItemDetail
+from stac_api.views.item import ItemsList
 from stac_api.views.views import AssetDetail
 from stac_api.views.views import AssetsList
 from stac_api.views.views import AssetUploadAbort
@@ -21,8 +23,6 @@ from stac_api.views.views import CollectionAssetUploadDetail
 from stac_api.views.views import CollectionAssetUploadPartsList
 from stac_api.views.views import CollectionAssetUploadsList
 from stac_api.views.views import ConformancePageDetail
-from stac_api.views.views import ItemDetail
-from stac_api.views.views import ItemsList
 from stac_api.views.views import LandingPageDetail
 from stac_api.views.views import SearchList
 from stac_api.views.views import recalculate_extent
